* Performance Update - add support for multiple command queues - provide minimal clgemm and clgemv functions for better performance - add support for pointer offsets as arguments - use C++ maps to cache frequently used memory queries between host and device memory
* Code Update
- disable OpenCLMemory copy constructor
- add support for shared_ptr

+++ include/caffe/util/im2col.hpp
@@ -22,14 +22,18 @@ void im2col_gpu(const Dtype* data_im, const int channels,
     const int stride_w, Dtype* data_col);
 
 template<typename Dtype>
-void im2col_gpu(const Dtype* data_im, const int data_im_step, const int channels, const int height, const int width,
+void im2col_gpu(
+    const Dtype* data_im, const size_t data_im_step,
+    const int channels, const int height, const int width,
 		const int kernel_h, const int kernel_w,
 		const int pad_h, const int pad_w,
 		const int stride_h, const int stride_w,
-		Dtype* data_col, const int data_col_step);
+		Dtype* data_col, const size_t data_col_step);
 
 template<typename Dtype>
-void im2col_gpu(const Dtype* data_im, const int bottom_step, const int num_images, const int num_channels, const int height, const int width,
+void im2col_gpu(
+    const Dtype* data_im, const int bottom_step,
+    const int num_images, const int num_channels, const int height, const int width,
 		const int kernel_h, const int kernel_w,
 		const int pad_h, const int pad_w,
 		const int stride_h, const int stride_w,
